+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/detection_engine/alert_suppression/eql_rule.cy.ts
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-import { getEqlSequenceRule } from '../../../../objects/rule';
+import { getEqlRule } from '../../../../objects/rule';
 
 import { login } from '../../../../tasks/login';
 import { visit } from '../../../../tasks/navigation';
@@ -34,24 +34,18 @@ import {
 
 const SUPPRESS_BY_FIELDS = ['agent.type'];
 
+// Skip in MKI due to flake
 describe(
-  'Detection Rule Creation - EQL Rules - With Alert Suppression',
+  'Detection EQL Rules - Alert suppression',
   {
-    // skipped in MKI as it depends on feature flag alertSuppressionForEsqlRuleEnabled
-    // alertSuppressionForEsqlRuleEnabled feature flag is also enabled in a global config
     tags: ['@ess', '@skipInServerlessMKI'],
-    env: {
-      kbnServerArgs: [
-        `--xpack.securitySolution.enableExperimental=${JSON.stringify([
-          'alertSuppressionForSequenceEqlRuleEnabled',
-        ])}`,
-      ],
-    },
   },
   () => {
-    describe('with sequence queries ', () => {
-      const rule = getEqlSequenceRule();
-
+    describe('with non-sequence queries', () => {
+      const rule = getEqlRule();
+      before(() => {
+        cy.task('esArchiverLoad', { archiveName: 'auditbeat_multiple' });
+      });
       beforeEach(() => {
         deleteAlertsAndRules();
         login();
@@ -59,6 +53,9 @@ describe(
 
         selectEqlRuleType();
         fillDefineEqlRule(rule);
+      });
+      after(() => {
+        cy.task('esArchiverUnload', { archiveName: 'auditbeat_multiple' });
       });
 
       it('creates a rule with a "per rule execution" suppression duration', () => {
