+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/detection_engine/alert_suppression/essentials_license_check_rule_form.cy.ts
@@ -21,7 +21,7 @@ import {
 import { CREATE_RULE_URL } from '../../../../urls/navigation';
 
 describe(
-  'Detection rules, Alert Suppression for Essentials tier',
+  'Alert Suppression - Essentials tier license check',
   {
     tags: ['@serverless'],
     env: {
