+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/detection_engine/alert_suppression/indicator_match_rule.cy.ts
@@ -36,7 +36,7 @@ import { deleteAlertsAndRules } from '../../../../tasks/api_calls/common';
 const SUPPRESS_BY_FIELDS = ['myhash.mysha256', 'source.ip.keyword'];
 
 describe(
-  'Detection rules, Indicator Match, Alert Suppression',
+  'Indicator Match - Alert suppression',
   {
     tags: ['@ess', '@serverless'],
   },
