+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/detection_engine/alert_suppression/machine_learning_rule.cy.ts
@@ -40,7 +40,7 @@ import { getDetails } from '../../../../tasks/rule_details';
 import { CREATE_RULE_URL } from '../../../../urls/navigation';
 
 describe(
-  'Machine Learning Detection Rules - Creation',
+  'Machine Learning Detection Rules - Alert suppression',
   {
     tags: ['@ess', '@serverless'],
   },
