+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/detection_engine/rule_creation/esql_rule.cy.ts
@@ -16,10 +16,6 @@ import {
   RULE_NAME_HEADER,
   RULE_TYPE_DETAILS,
   RULE_NAME_OVERRIDE_DETAILS,
-  DEFINITION_DETAILS,
-  SUPPRESS_BY_DETAILS,
-  SUPPRESS_FOR_DETAILS,
-  SUPPRESS_MISSING_FIELD,
 } from '../../../../screens/rule_details';
 
 import { ESQL_QUERY_BAR } from '../../../../screens/create_new_rule';
@@ -41,14 +37,6 @@ import {
   fillRuleName,
   fillDescription,
   getAboutContinueButton,
-  fillAlertSuppressionFields,
-  selectAlertSuppressionPerInterval,
-  setAlertSuppressionDuration,
-  selectDoNotSuppressForMissingFields,
-  continueFromDefineStep,
-  fillAboutRuleMinimumAndContinue,
-  skipScheduleRuleAction,
-  interceptEsqlQueryFieldsRequest,
   createRuleWithNonBlockingErrors,
 } from '../../../../tasks/create_new_rule';
 import { login } from '../../../../tasks/login';
@@ -68,7 +56,7 @@ const workaroundForResizeObserver = () =>
   });
 
 describe(
-  'Detection ES|QL rules, creation',
+  'Detection ES|QL rules - Rule Creation',
   {
     tags: ['@ess', '@serverless', '@skipInServerlessMKI'],
   },
@@ -239,61 +227,6 @@ describe(
         createRuleWithoutEnabling();
 
         cy.get(INVESTIGATION_FIELDS_VALUE_ITEM).should('have.text', CUSTOM_ESQL_FIELD);
-      });
-    });
-
-    describe('Alert suppression', () => {
-      beforeEach(() => {
-        login();
-        visit(CREATE_RULE_URL);
-      });
-      it('shows custom ES|QL field in investigation fields autocomplete and saves it in rule', function () {
-        const CUSTOM_ESQL_FIELD = '_custom_agent_name';
-        const SUPPRESS_BY_FIELDS = [CUSTOM_ESQL_FIELD, 'agent.type'];
-
-        const queryWithCustomFields = [
-          `from auditbeat* metadata _id, _version, _index`,
-          `eval ${CUSTOM_ESQL_FIELD} = agent.name`,
-          `drop agent.*`,
-        ].join(' | ');
-
-        workaroundForResizeObserver();
-
-        selectEsqlRuleType();
-
-        interceptEsqlQueryFieldsRequest(queryWithCustomFields, 'esqlSuppressionFieldsRequest');
-        fillEsqlQueryBar(queryWithCustomFields);
-
-        cy.wait('@esqlSuppressionFieldsRequest');
-        fillAlertSuppressionFields(SUPPRESS_BY_FIELDS);
-        selectAlertSuppressionPerInterval();
-        setAlertSuppressionDuration(2, 'h');
-        selectDoNotSuppressForMissingFields();
-        continueFromDefineStep();
-
-        // ensures details preview works correctly
-        cy.get(DEFINITION_DETAILS).within(() => {
-          getDetails(SUPPRESS_BY_DETAILS).should('have.text', SUPPRESS_BY_FIELDS.join(''));
-          getDetails(SUPPRESS_FOR_DETAILS).should('have.text', '2h');
-          getDetails(SUPPRESS_MISSING_FIELD).should(
-            'have.text',
-            'Do not suppress alerts for events with missing fields'
-          );
-        });
-
-        fillAboutRuleMinimumAndContinue(rule);
-        skipScheduleRuleAction();
-        createRuleWithoutEnabling();
-
-        // ensures rule details displayed correctly after rule created
-        cy.get(DEFINITION_DETAILS).within(() => {
-          getDetails(SUPPRESS_BY_DETAILS).should('have.text', SUPPRESS_BY_FIELDS.join(''));
-          getDetails(SUPPRESS_FOR_DETAILS).should('have.text', '2h');
-          getDetails(SUPPRESS_MISSING_FIELD).should(
-            'have.text',
-            'Do not suppress alerts for events with missing fields'
-          );
-        });
       });
     });
   }
